+++ jest.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+    testPathIgnorePatterns : ["/node_modules/", "/thespoon/", "/mobileApp/"]
+};
