+++ package.json
@@ -4,7 +4,7 @@
   "description": "The server application of the project",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest --verbose",
     "start": "node server.js",
     "heroku-postbuild": "cd thespoon && npm install --only=dev && npm install && npm run build"
   },
@@ -26,5 +26,9 @@
     "sequelize": "^5.21.2",
     "swagger-ui-express": "^4.1.2",
     "yamljs": "^0.3.0"
+  },
+  "devDependencies": {
+    "jest": "^24.9.0",
+    "supertest": "^4.0.2"
   }
 }
